+++ app/src/main/java/com/sedilant/yambol/domain/DeleteTeamObjectiveUseCase.kt
@@ -2,6 +2,6 @@ package com.sedilant.yambol.domain
 
 interface DeleteTeamObjectiveUseCase {
     suspend operator fun invoke(
-        objectiveId: Int, description: String, isFinish: Boolean, teamId: Int
+        objectiveId: Int, description: String, isFinish: Boolean, teamId: Long
     )
 }
