+++ app/src/main/java/com/sedilant/yambol/domain/get/GetPlayerByIdUseCase.kt
@@ -2,6 +2,6 @@ package com.sedilant.yambol.domain.get
 
 import com.sedilant.yambol.ui.home.models.PlayerUiModel
 
-interface GetPlayerByIdUseCase{
-    suspend operator fun invoke (id: Int): PlayerUiModel
+interface GetPlayerByIdUseCase {
+    suspend operator fun invoke(id: Long): PlayerUiModel
 }
