+++ app/src/main/java/com/sedilant/yambol/domain/insert/CreateTrainUseCaseImpl.kt
@@ -1,14 +1,14 @@
 package com.sedilant.yambol.domain.insert
 
-import com.sedilant.yambol.data.TeamRepository
-import com.sedilant.yambol.data.entities.TrainEntity
+import com.sedilant.yambol.data.team.TeamRepository
+import com.sedilant.yambol.data.team.TrainEntity
 import java.util.Date
 import javax.inject.Inject
 
 class CreateTrainUseCaseImpl @Inject constructor(
     private val teamRepository: TeamRepository
 ) : CreateTrainUseCase {
-    override suspend fun invoke(date: Date, time: Float, concepts: List<String>, teamId: Int): Int {
+    override suspend fun invoke(date: Date, time: Float, concepts: List<String>, teamId: Long): Long {
         return teamRepository.insertTrain(
             TrainEntity(
                 date = date.time,
